🐛 Somewhat failed attempt to improve error handling in login process; clarify error messages, remove unused code in auth config module, and move authorize functions to seperate file
